implement analyze and output samples to .dartdoc_test directory
